@@ -36,7 +36,8 @@ dependencies:
   cupertino_icons: ^1.0.8
   device_preview: ^1.2.0
   image_picker: ^1.1.2
-  google_maps_flutter: ^2.10.1
+  google_maps_flutter: ^2.11.0
+
   intl: ^0.19.0
   firebase_core: ^2.25.0
   firebase_auth: ^4.17.0
@@ -49,6 +50,8 @@ dependencies:
   http: ^1.2.0
   # Added for permission handling
   permission_handler: ^11.3.0
+  geocoding: ^3.0.0
+  flutter_svg: ^2.0.17
 
 dev_dependencies:
   flutter_test:
