@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:device_preview/device_preview.dart';
 import 'package:firebase_core/firebase_core.dart';
+
 import 'dashboarduser.dart';
 import 'firebase_options.dart';
 
@@ -30,7 +31,7 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.green),
         useMaterial3: true,
       ),
-      home: const EcoBinDashboard(), // Open Get Started Page first
+      home: EcoBinDashboard(), // Open Dashboard Page
       builder: DevicePreview.appBuilder,
       locale: DevicePreview.locale(context),
     );
